requirements.txt
PyYAML~=5.4.1
selenium~=3.141.0
Appium-Python-Client~=1.1.0
pytest~=6.2.2